pyproject.toml
[tool.poetry]
name = "tiff"
version = "0.1.0"
description = "GeoTiff utils package"
authors = ["Moleque <molecada@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.19.5"
rasterio = "^1.2.3"
pyproj = "^3.0.1"
pillow = "^8.2.0"
geopy = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
